requirements.txt
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
cssbeautifier==1.15.4
djlint==1.36.4
dnspython==2.8.0
EditorConfig==0.17.1
email-validator==2.3.0
Faker==40.8.0
Flask==3.1.3
Flask-SQLAlchemy==3.1.1
greenlet==3.3.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jsbeautifier==1.15.4
json5==0.13.0
MarkupSafe==3.0.3
pathspec==1.0.4
pycparser==3.0
PyMySQL==1.1.2
python-dotenv==1.2.2
PyYAML==6.0.3
regex==2026.2.28
requests==2.32.5
six==1.17.0
SQLAlchemy==2.0.48
tqdm==4.67.3
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.6
